% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{hamming}
\alias{hamming}
\title{Hamming distance matrix}
\usage{
hamming(x)
}
\arguments{
\item{x}{matrix with individuals in rows and markers in columns.}
}
\value{
A square symmetric matrix of Hamming distances.
}
\description{
Compute the pairwaise Hamming distances (no. of differences) between individuals.
Accepts any storage class with a defined unequality operator \code{!=} including integer (e.g., minisatellite repeat counts) or character (nucleotides).
\code{NA}s not allowed.
}
