% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{thd.adjust}
\alias{thd.adjust}
\title{Principal component-based correction for population structure}
\usage{
thd.adjust(x, H, m, method = "fdr", alpha = 0.05)
}
\arguments{
\item{x}{vector of THD values}

\item{H}{genetic distance matrix used to generate \code{x}}

\item{m}{maximum number of PCs to be considered. Typically, the no. of markers used to generate \code{H}.}

\item{method}{P-value adjustment method passed to \code{p.adjust}. Default = false discovery rate (FDR).}

\item{alpha}{significance threshold to retain a given PC.}
}
\value{
Matrix with \code{length(x)} rows and one column per PC significantly associated with \code{x}.
If no PC is found, a 0-column matrix is returned.
}
\description{
Generate a matrix of principal components (PCs) significantly associated with THD.
Include these PCs in multivariate models to correct for population structure.
}
