% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thd.R
\name{thd}
\alias{thd}
\title{Thresholding model predictions}
\usage{
thd(sdm, occ, type = "")
}
\arguments{
\item{sdm}{ENM suitability prediction}

\item{occ}{Occ of long and lat}

\item{type}{mtp,p10}
}
\description{
Thresholding model predictions
}
